chore: Remove test-core and align Tests
