+++ sdk/build.gradle.kts
@@ -70,6 +70,7 @@ dependencies {
     api(libs.kuksa.java.sdk)
     implementation(libs.kotlinx.coroutines.core)
 
+    implementation(libs.kotest)
     testImplementation(libs.junit.jupiter.api)
     testRuntimeOnly(libs.junit.jupiter.engine)
     testImplementation(libs.kotest.runner.junit5)
