--- test-core/src/main/java/org/eclipse/velocitas/test/TestResourceFile.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2023 - 2024 Contributors to the Eclipse Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- */
-
-package org.eclipse.velocitas.test
-
-import java.io.File
-
-private val classLoader: ClassLoader? = TestResourceFile::class.java.classLoader
-
-class TestResourceFile(path: String) : File(
-    classLoader!!.getResource(path)?.file ?: error("File does not exist: '$path'"),
-)
