--- test-core/src/main/java/org/eclipse/velocitas/test/extension/FloatExtension.kt
@@ -1,26 +1,0 @@
-/*
- * Copyright (c) 2023 - 2024 Contributors to the Eclipse Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- */
-
-package org.eclipse.velocitas.test.extension
-
-import kotlin.math.abs
-
-fun Float.equals(other: Float, epsilon: Float): Boolean {
-    return abs(this - other) <= epsilon
-}
