--- test-core/src/main/java/org/eclipse/velocitas/test/kotest/Tag.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- */
-
-package org.eclipse.velocitas.test.kotest
-
-import io.kotest.core.NamedTag
-
-val Integration = NamedTag("Integration")
-
-val InsecureDataBroker = NamedTag("InsecureDataBroker") // no tls, no authentication
-val Insecure = NamedTag("Insecure")
-
-val SecureDataBroker = NamedTag("SecureDataBroker")
-val Secure = NamedTag("Secure") // tls enabled, authentication enabled
-val Authentication = NamedTag("Authentication")
-val Tls = NamedTag("Tls")
-
-val Unit = NamedTag("Unit")
