+++ vss-processor-plugin/src/test/kotlin/org/eclipse/velocitas/TestResourceFile.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.eclipse.velocitas
+
+import java.io.File
+
+private val classLoader: ClassLoader? = TestResourceFile::class.java.classLoader
+
+class TestResourceFile(path: String) : File(
+    classLoader!!.getResource(path)?.file ?: error("File does not exist: '$path'"),
+)
